Concourse running Docker, Access to Vault over tls, Token auth.
Necessary definitions made to support tls communication with Vault:
a CA certificate available to Concourse.  This cert is placed in
a directory that is mounted in the container.  It can hold other
certs, etc. as needed.

The prototype structure reworked.  More private/local info now
in private (populated via the prototype files) instead of in
docker-compose.override.yml  The prototype/init script automates
populating private and bin directories.

The up and down scripts have a slightly more elaborate method of
collecting cred's and settings.

README-LOCAL.md expanded and edited.  It is converging on a single
README for all the config's supported.  Not there yet.